Fix from Tom- fixed capture logic
Fixed issue, in short, the DDR write controller needed to look at
data_avail from the FIFO and wr_ready from the DDR interface
separately. It was ANDing them together in the previous
version of the model which was causing the off-by-one error in
the write counter.

Try simulating this and look at the logic analyzer. You should
see that we capture 4 range swaths of 50 samples each for
200 samples total, then read this out of DDR with a frame size
of 64. The last frame will contain garbage data after the 8th
sample because the capture length is not an even multiple of
the read frame size. This shows that we can handle
independent write and read frame sizes correctly. In
hardware, you might set the capture length to something like
500k and read it out in smaller chunks of 1k or so.