Added Validator abstract class, properties to Radar_pl_configuration
Validator abstract class has method
allPropertiesAreSingletonAndDefined(obj) which makes it
easy for subclasses to validate their properties.

A later add on would be to add optional arguments to
specify which properties are tested, the options for
inclusion or exclusion lists for the test.

Radar_pl_configuration has a bare minimum of properties,
currently not including NCO/frequency step sizes.
Frequency related properties are needed.

Additionally, current  validation is limited to ensuring
that all properties are defined, but none of their
mathematical relationships are currently tested for
validity.