Moved RadarSetup output properties to methods to simplify, draft stage.
Created Validator.getDiscriminator(typeString,propertyCellArray)
which returns a function that returns a boolean on
whether to accept a given property string. Used to
give the allPropertiesAreSingletonAndDefined and
allPropertiesAreDefined a whitelist and blacklist mode.

In the midst of adding similar functionality to
parseConstructorInputForClassProperties()
Function no longer takes variable inputs separated
as varargin{:} but rather just varargin, which means
the values come as a single cell array argument,
rather than as a variable number of arguments.
The argument varargin has been renamed
vararginCellArray, since it is actually only one
argument now. This allows vararginCellArray to be
the first argument, rather than the last, and can now
be followed by optional arguments to implement
whitelist/blacklist property behavior.

Drafting methods:
getRadarPerformance()
isPulseTimingValid()
getRadarPlConfig()