Added new AXI4 register names
Changed the AXI4 write register names to avoid conflicts with matlab workspace variables and so they are easily identified in the .m files. Could have just changed the workspace variable names but this was more complicated due to multiple files. Felt changing all the register names would be clearer in the long run. Also the read script had a reference to the TriggerCapture register that was removed in early versions before CPI logic was added. Verified simulation and operation of the bitstream with this version.