Added NCO and Chirp logic from Simple_pulse_generator
Brought in the NCO and Chirp generation logic, modified the system controller to provide needed signals, updated some of the AXI4 write register names, updated init script to make necessary calculations. A quick look at the simulation looks correct for this version. Note that I was not able to loop DAC data back into the ADC as we could in Simple_pulse_generator, not sure why. Saved that version in a separate commit.