Add N for accumulator, start and end chirp steps, organizing init_ADC
Preparing more organized initialization and configuration.
Moving in from several directions.
-init_ADC_Capture_4x4_IQ_DDR4.m modified to clear
variables which aren't needed to run the model in
simulation.  A cell array of all requiredVars which are not
cleared is provided.
Note that this must be tested on both the simulation
and the realtime script ADC_IO_Capture_DDR4_Read.m
to be sure that all required variables are available.
The intent of this process is to discover all variables that
are actually required for operation, and to
a) Ensure an analog exists within the new config classes
b) Add any new variables required by the radar.

-RadarSetup has an update to getRadarPlConfig, but this
is not complete and accurate yet.  It serves as a proof of
concept so far.

-Considering new names
behavior -> pl_config
Radar_pl_configuration -> pl_register_config for inputs
external to the programmable logic.
RFSoC_Clock_Settings -> pl_synthesis_config for values
which are baked into the bitstream.

-getConfig.simulation and getConfig.realtime could be a
way to store initialization information.