Added sin HDL Coder block to look at pipelining in complex IP functions
Added a divide on the output of HDL counter in the NCO section to generate a value between 0 and 1 to feed a sin block I added so I could look at various options for pipelining in the IP block. Need to follow up with Tom on this to understand why I cannot get the simulator to show additional latency. Without pipelining it seems like we would be limited to very clock rates based on the pipeline steps in the native Xilinx and Intel IP. Follow up with Tom to see what I'm missing in the configuration, the documentation describes how to implement pipelining for this block.