Fixed capture length and updated specAn operating parameters
In the process of combining sections from various designs we ended up with CaptureLength and CaptureSize both representing the same parameter. Chnaged code to use CaptureLength only. Changed range swath length so that we could capture then entire pulse, could have also used range gate delay to move the window but just made the range swath 25% bigger than the pulse width so that we'll have a buffer. Specified  specAn parameters for resolution method, window length and windowing mode so we get a nice view of the frequency spectrum