Reduced Trigger Capture delay time
The simulation was initially setup to initialize the DDR4 model and have a significant delay before the first pulse went out in the simulation. Depending on the configuration of the pulse width, range gate and other parameters that affect the DDR4 capture length this could be significant. Removed the option to initialize the DDR4 Model and changed this delay in the top level simulation to 10.