Fixed exist bug, convert scalar to bool with == 1
It seems like the exist() method should provide a boolean,
but it returns a double, which isn't automagically
converted to a boolean within the assert.
To deal with this, we just compare to 1 to get the
desired behavior.

At present, setupOnce just verifies that all the required
variable are present when you run the original
init_ADC_Capture_4x4_IQ_DDR4.m script, which is
pretty much guaranteed to work.

testInitOne() runs the object oriented initialization
with some static parameters, but makes no test
assertions.