Add initializeObjects(), prep comparison in initializeScript,
initialize.initializeObjects() is a function that returns a
structure containing the 3 pl_config objects,
synthesisConfig, radarSetup, and pl_register_config.
These are intended to provide all the data required to
synthesize, simulate, and initialize axi4 registers at
runtime.  This covers both
ADC_Capture_4x4_IQ_DDR4.slx and
ADC_IO_Capture_DDR4_Read.m

initializeObjects() gives the opportunity to verify that
the object initialization is giving all the required types
of parameters, but does not verify that the input to
output calculation is equivalent between the original
script and the new objects.

initialize.initializeScript has been updated to contain
both a function based initialization and a object
oriented initialization.  The input parameters of the
object oriented init are captured as a cell array and
converted to structured arrays to aid future comparison
of the input to output conversion between the
function and object based initializations.
This comparison has not yet been instrumented or
performed, and will require conforming the function
init input interface to the object input parameters, and the
object output parameters to the function output
parameters.

This script is setup to perform much the same experiment as was being developed in the test package.