init: All required vars added in comments to setup oop initialization
All required variable are listed below line 126.
There is a block of properties on line 141 which
should be absorbed into Mask settings.

The DDR properties on line 153 could remain, but
they should be absorbed into a single struct.

The simulation specific parameters are understandable, but
should be implemented differently (not as a global).
A simulation vs hardware flag could be added to
the initialize.initializeObjects function, or it could be setup
as a Static class with hardware() and simulation() specific methods.
In any case, the overall interface should be maintained, rather
than adding another global variable.