Add CoutDir to Mask, Replace references to 'N' parameter in model
Added CountDir boolean to mask dialog box for Modified HDL Counter
to improve re-usability.

Replaced all references to base workspace variable 'N'
throughout Simulink model to
initObjects.synthesisConfig.N_accumulator
This is an initial test of the compatibility of Matlab object parameters
with Simulink.  Doing this cleanly required adding a simulink Fcn
block to allow the longer variable name without obscuring the math.

Named several top level constant blocks to make the
Simulink Model Explorer easier to use.

Pruned the init m-file a bit

Note:
Though at this point most of the required parameters are defined
by the OOP initialization, the values are different in many cases,
and the calculations still require validation.
For example, the Read.m script won't initialize with the current
parameters because of the ratio between CaptureLength and
CPILength.