data type mismatched ddr4 read and write using psps
Two notes:
The write dataLength (line 10) and the CaptureLength (line 11)
should not be equal... but the correct ratio is unknown.

The received data matches when the input is zero, but is
scrambled for other inputs.  It is apparent that the datatype
doesn't match.
Needs to be casted or converted either before write or after
read, or both.