slx: completes hdlworkflow for RTL Generation
Removed MaxValue input for Modified HDL Counter, and
folded the calculation into the mask initialization.
Previous implementation used a depreciated 'Fcn' block, which 
does support HDL generation, but only for native floating point.
In reality, the Fcn block was only being used to calculate the value of a constant, something that no PL resources should 
be used for.  With the calculation now in the mask initialization, 
I'm hopeful that it is optimized away, but not certain.