Unused controller m-files moved to unused folder
By moving unused m-files to a folder called unused within
the +controller package, it prevents a person from wasting
time searching for where they are being used.

The folder is not on the path and is not a package, so
it gives the observer some confidence that it is not
being utilized at all and can safely be ignored.